🧩 GX777N task: keep insights diagnostics below hotspot limit
